feat: 优化面板
